# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniCare Policyholder Assistant",
    "description": "Coverage answers grounded in policy documents with section citations, claim lookup, and claim submission with explicit confirmation.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "description": "Liveness. Returns exactly {\"status\": \"healthy\"} - the graded shape.",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/health/deep": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Deep",
        "description": "Readiness. Used by compose and by the frontend to decide whether to\noffer the voice button at all.",
        "operationId": "health_deep_api_v1_health_deep_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeepHealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "description": "Answer one turn.\n\nCoverage questions are answered from the policy documents with citations;\nclaim questions call the claims tools. Filing a claim returns a\nconfirmation prompt first - reply \"yes\" in the next message to proceed.",
        "operationId": "chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice/token": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Voice Token",
        "description": "Mint a short-lived LiveKit JWT scoped to one room.\n\nServer-side only: the API secret never reaches the browser.",
        "operationId": "voice_token_api_v1_voice_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations/{user_id}": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "List Conversations",
        "description": "A user's conversations, most recent first.\n\nOwned by the gateway, so history stays readable while the agent is down.",
        "operationId": "list_conversations_api_v1_conversations__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Conversations Api V1 Conversations  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations/{conversation_id}/messages": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Conversation Messages",
        "description": "Full history with sources and tool calls, for rehydrating the UI.",
        "operationId": "conversation_messages_api_v1_conversations__conversation_id__messages_get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Conversation Messages Api V1 Conversations  Conversation Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_\\-]+$",
            "title": "User Id"
          },
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "channel": {
            "type": "string",
            "enum": [
              "text",
              "voice"
            ],
            "title": "Channel",
            "default": "text"
          },
          "stt_confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stt Confidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "message"
        ],
        "title": "ChatRequest",
        "examples": [
          {
            "message": "A pipe burst in my kitchen. Am I covered?",
            "user_id": "usr_123"
          },
          {
            "message": "What is the status of claim CLM-8821?",
            "user_id": "usr_123"
          }
        ]
      },
      "ChatResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolCall"
            },
            "type": "array",
            "title": "Tool Calls"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "response",
          "conversation_id"
        ],
        "title": "ChatResponse",
        "examples": [
          {
            "conversation_id": "cnv_01HZ",
            "response": "Yes - sudden pipe bursts are covered up to $25,000 with a $500 deductible.",
            "sources": [
              "sample_policy.md \u00a7 Section 1: Home Water Damage Coverage"
            ],
            "tool_calls": [
              {
                "arguments": {
                  "query": "burst pipe water damage coverage"
                },
                "name": "search_policy_documents",
                "status": "ok"
              }
            ]
          }
        ]
      },
      "DeepHealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded"
            ],
            "title": "Status",
            "default": "healthy"
          },
          "dependencies": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "up",
                "down"
              ]
            },
            "type": "object",
            "title": "Dependencies"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DeepHealthResponse",
        "description": "``GET /api/v1/health/deep`` - readiness, used by compose and the UI."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded"
            ],
            "title": "Status",
            "default": "healthy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "HealthResponse",
        "description": "``GET /api/v1/health`` - serializes to exactly {\"status\": \"healthy\"}."
      },
      "ToolCall": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "error",
              "awaiting_confirmation",
              "denied"
            ],
            "title": "Status",
            "default": "ok"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ToolCall",
        "description": "One tool invocation, surfaced in the response for auditability."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoiceTokenRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_\\-]+$",
            "title": "User Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "VoiceTokenRequest"
      },
      "VoiceTokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "examples": [
              "ws://localhost:7880"
            ]
          },
          "room": {
            "type": "string",
            "title": "Room"
          },
          "expires_in": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 60.0,
            "title": "Expires In",
            "default": 900
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "url",
          "room"
        ],
        "title": "VoiceTokenResponse"
      }
    }
  }
}